[#132] Flyway database migrations (#140)
